refactor(test): Split tests into separate modules
Move version, basic, and action_script tests from bin.rs into
src/tests/ (version.rs, basic.rs, action_script.rs).  bin.rs keeps
only main and test dispatch.  Matches the structure used on
add-orphan-pts-console-socket-callback for consistency.
