+++ src/tests/mod.rs
@@ -1,0 +1,5 @@
+//! Integration tests: each test in its own module.
+
+pub mod action_script;
+pub mod basic;
+pub mod version;
